+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## [0.0.1] - 2020-08-08
 ### Updates
 - Update icon launcher
+- Added forfeit button
+- Added undo button
+- Added player turn
+
+### Bugs
+- Fixed draw when on board edge
+- Fixed king capturing defender pieces
 
 ## [0.0.0] - 2020-08-03
 ### Updates